+++ cli/src/lib.rs
@@ -30,6 +30,7 @@ mod event_formatter;
 mod parser;
 mod short_id;
 mod table;
+mod todo_editor;
 mod todo_formatter;
 
 pub use crate::{